refactor(dsl): move workflow init activities out of validation
