+++ tracecat/dsl/worker.py
@@ -21,12 +21,12 @@ with workflow.unsafe.imports_passed_through():
     from tracecat import config
     from tracecat.dsl.action import DSLActivities
     from tracecat.dsl.client import get_temporal_client
-    from tracecat.dsl.interceptor import SentryInterceptor
-    from tracecat.dsl.plugins import TracecatPydanticAIPlugin
-    from tracecat.dsl.validation import (
+    from tracecat.dsl.init_activities import (
         resolve_time_anchor_activity,
         resolve_workflow_concurrency_limits_enabled_activity,
     )
+    from tracecat.dsl.interceptor import SentryInterceptor
+    from tracecat.dsl.plugins import TracecatPydanticAIPlugin
     from tracecat.dsl.workflow import DSLWorkflow
     from tracecat.ee.interactions.service import InteractionService
     from tracecat.logger import logger
